refactor: my page를 서버 컴포넌트로 변경
