+++ src/app/(iTracker)/my/layout.tsx
@@ -1,7 +1,19 @@
+'use client';
+
+import { isLogin } from '@/features/auth/api/oauth';
 import { FixedBottomNavbar } from '@/shared/components/FixedBottomNavbar';
-import React from 'react';
+import { useRouter } from 'next/navigation';
+import React, { useEffect } from 'react';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!isLogin()) {
+      router.push('/login?needLogin=true');
+    }
+  }, [router]);
+
   return (
     <div className="flex-1 w-full flex flex-col">
       {children}
