+++ src/app/(iTracker)/oauth/redirected/kakao/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import KakaoCallback from '@/features/auth/components/KakaoCallback';
 import { Loading } from '@/shared/components/Loading';
 import { Suspense } from 'react';
